// src/app/components/models/estoque.model.ts
export class Estoque {
  id?: string;
  idProduto?: string;
  nomeProduto?: string;
  codBarras?: string;
  dataEstoque?: string;
  quantidade?: number;
  custo?: number;
  dataValidade?: string;
  taxaShopee?: number;
  valorVendaSugerido?: number;
  lucroReais?: number;
  lucroPercentual?: number;
  dataCriacao?: string;
  dataAtualizacao?: string;
}